requirements.txt
amqp==5.0.9
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
autopep8==1.6.0
backports.zoneinfo==0.2.1
bcrypt==3.2.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==6.3.1
Deprecated==1.2.13
distlib==0.3.4
dj-database-url==0.5.0
Django==3.2.12
django-braces==1.15.0
django-celery-beat==2.2.1
django-dotenv==1.4.2
django-filter==21.1
django-mailjet==0.3.1
django-timezone-field==4.2.3
djangorestframework==3.13.1
filelock==3.4.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.1
kombu==5.2.3
mailjet-rest==1.3.4
Markdown==3.3.6
mysqlclient==2.1.0
packaging==21.3
Pillow==9.0.0
pipenv==2022.1.8
platformdirs==2.5.0
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pyparsing==3.0.7
pypi==2.1
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
redis==4.1.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.8
validate-email==1.3
vine==5.0.0
virtualenv==20.13.1
virtualenv-clone==0.5.7
wcwidth==0.2.5
whitenoise==6.0.0
wrapt==1.13.3
zipp==3.7.0
